refactor: move and extract methods from Program
