+++ src/ScriptCs/ScriptCs.csproj
@@ -82,6 +82,7 @@
     <Compile Include="ConfigMask.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="ProfileOptimizationShim.cs" />
+    <Compile Include="Application.cs" />
     <Compile Include="VersionWriter.cs" />
     <Compile Include="Command\CleanCommand.cs" />
     <Compile Include="Command\CommandFactory.cs" />
@@ -116,7 +117,7 @@
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
     </None>
     <None Include="scriptcs">
-    	<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
   <ItemGroup>
